pyproject.toml
[project]
name = "pet-eff-mob"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=5.0.0",
    "django>=5.2.6",
    "djangorestframework>=3.16.1",
    "psycopg2-binary>=2.9.10",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
]
